refactor: Move product card container to templates
