+++ src/components/Templates/Home/Home.component.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 import axios from 'axios';
 import { IProduct } from '../../../types/models/Product';
-import { ProductCardContainer } from '../../Molucules/ProductCardConatiner';
+import { ProductCardContainer } from '../ProductCardConatiner';
 import { Box } from '@mui/material';
 import { Categories } from '../../Molucules/Categories';
